refactor: 객체 중심으로 패키지 정리 -Product, User, Wish - 하위에 service, controller repository, domain 패키지
