+++ src/main/java/gift/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package gift.repository;
+package gift.product.repository;
 
-import gift.model.Product;
+import gift.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
