+++ src/main/java/gift/user/domain/User.java
@@ -1,5 +1,6 @@
-package gift.model;
+package gift.user.domain;
 
+import gift.wish.domain.WishlistItem;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
