+++ src/main/java/gift/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package gift.repository;
+package gift.user.repository;
 
-import gift.model.User;
+import gift.user.domain.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
