+++ src/main/java/gift/wish/domain/WishlistDTO.java
@@ -1,4 +1,4 @@
-package gift.model;
+package gift.wish.domain;
 
 public class WishlistDTO {
     Long userId;
